+++ extension/src/DataSource/DataExtractionProvider/index.ts
@@ -1,12 +1,6 @@
 export * from "./DataExtractionProvider";
 
-import { ComposedDataExtractionProviderFactory } from "./ComposedDataExtractionProviderFactory";
-import { JsDataExtractionProviderFactory } from "./JsDataExtractionProviderFactory";
-import { GenericDataExtractionProviderFactory } from "./GenericDataExtractionProviderFactory";
-
-export const defaultDataExtractionProviderFactory = new ComposedDataExtractionProviderFactory(
-	[
-		new JsDataExtractionProviderFactory(),
-		new GenericDataExtractionProviderFactory(),
-	]
-);
+export { ComposedDataExtractionProviderFactory } from "./ComposedDataExtractionProviderFactory";
+export { JsDataExtractionProviderFactory } from "./JsDataExtractionProviderFactory";
+export { GenericDataExtractionProviderFactory } from "./GenericDataExtractionProviderFactory";
+export { ConfiguredDataExtractionProviderFactory } from "./ConfiguredDataExtractionProviderFactory";