+++ extension/src/DataSource/index.ts
@@ -1,1 +1,1 @@
-export * from "./Sources";
+export * from "./DataSourceImpl";